refactor: removed session from resear jwt token
